fix: resolve clippy errors and add workspace lint configuration
- rename from_str → parse_str on queue types to avoid FromStr trait
  confusion (ItemType, QueueOperation, QueueStatus, DestinationStatus)
- fix Range::contains, redundant closure, derivable Default impls
- add [workspace.lints.clippy] to suppress style lints that require
  larger refactoring (too_many_arguments, type_complexity, etc.)
- add [lints] workspace = true to all member crates
- update CI to check --lib --bins --tests (skip benches)

+++ src/rust/daemon/memexd/src/queue_init.rs
@@ -198,7 +198,7 @@ pub async fn initialize(
     info!("Initializing queue processor...");
 
     let processor_config = ProcessorConfig {
-        batch_size: config.queue_batch_size.unwrap_or(10) as i32,
+        batch_size: config.queue_batch_size.unwrap_or(10),
         poll_interval_ms: config.queue_poll_interval_ms.unwrap_or(500),
         worker_count: config.queue_worker_count.unwrap_or(4),
         backpressure_threshold: config.queue_backpressure_threshold.unwrap_or(1000),
